# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_client_upload_file_api_v1_pipeline__order_id__client_upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_client_upload_file_api_v1_pipeline__order_id__client_upload_post",
        "type": "object"
      },
...
      "FileCategory": {
        "description": "Категории загружаемых файлов.\n\nЕдиный канонический формат `.value` — **snake_case lowercase**\n(фаза B2.a 2026-04-21, B2.b 2026-04-22). До B2 значения\n`BALANCE_ACT`/`CONNECTION_PLAN` хранились в UPPER_CASE; переведены в\nlowercase. С B2.b совместимый shim `_missing_` удалён — API теперь\nотвергает uppercase-варианты с **422 Unprocessable Entity**.\n\nВ PostgreSQL тип `file_category` — метки совпадают с **именами** членов\nPython (`TU`, `BALANCE_ACT`, …), а не с `.value`. SQLAlchemy без\n`values_callable` persist имена, поэтому смена `.value` не требует\nAlembic-миграции enum. Значения `.value` используются для API,\nquery-параметров и сегментов пути в хранилище\n(`upload_dir/<order_id>/<category.value>/...`).",
        "enum": [
          "tu",
          "balance_act",
          "connection_plan",
          "heat_point_plan",
          "heat_scheme",
          "generated_excel",
          "generated_project",
          "other",
          "company_card",
          "signed_contract",
          "contract",
          "invoice",
          "final_invoice",
          "rso_scan",
          "rso_remarks"
        ],
        "title": "FileCategory",
        "type": "string"
      },
      "FileResponse": {
        "description": "Информация о загруженном файле.",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FileCategory"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          }
        },
        "required": [
          "id",
          "category",
          "original_filename",
          "content_type",
          "file_size",
          "created_at"
        ],
        "title": "FileResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipelineResponse": {
        "description": "Ответ эндпоинтов пайплайна и публичного submit на лендинге.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "required": [
          "message",
          "order_id"
        ],
        "title": "PipelineResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Admin-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Сервис автоматизированного проектирования узлов учёта тепловой энергии",
    "title": "УУТЭ Проектировщик",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pipeline/{order_id}/approve": {
      "post": {
        "description": "Инженерский approve: `advance_paid` → отправка проекта клиенту.",
        "operationId": "approve_project_api_v1_pipeline__order_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "_k",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": " K"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Approve Project",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/api/v1/pipeline/{order_id}/client-upload": {
      "post": {
        "description": "Загрузка файла клиентом через страницу upload-page.\n\nОтличие от /orders/{id}/files: после загрузки проверяет,\nнужно ли ещё что-то, и обновляет стейт-машину.",
        "operationId": "client_upload_file_api_v1_pipeline__order_id__client_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "description": "Категория файла",
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FileCategory",
              "description": "Категория файла"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_client_upload_file_api_v1_pipeline__order_id__client_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Client Upload File",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/api/v1/pipeline/{order_id}/client-upload-done": {
      "post": {
        "description": "Клиент завершил загрузку файлов.\n\nКнопка «Готово» на странице upload-page.\nЗапускает проверку полноты данных.",
        "operationId": "client_upload_done_api_v1_pipeline__order_id__client_upload_done_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Client Upload Done",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/api/v1/pipeline/{order_id}/confirm-advance": {
      "post": {
        "description": "Инженер подтверждает получение аванса.\n\nДоступно только в статусе contract_sent.\nЗапускает process_advance_payment — перевод в advance_paid.",
        "operationId": "confirm_advance_payment_api_v1_pipeline__order_id__confirm_advance_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "_k",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": " K"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Confirm Advance Payment",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/api/v1/pipeline/{order_id}/confirm-final": {
      "post": {
        "description": "Инженер подтверждает получение остатка оплаты.\n\nДоступно только в статусе awaiting_final_payment.\nПереводит заявку в completed.",
        "operationId": "confirm_final_payment_api_v1_pipeline__order_id__confirm_final_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "_k",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": " K"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Confirm Final Payment",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/api/v1/pipeline/{order_id}/resend-corrected-project": {
      "post": {
        "description": "Инженер повторно отправляет исправленный проект клиенту.",
        "operationId": "resend_corrected_project_endpoint_api_v1_pipeline__order_id__resend_corrected_project_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "_k",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": " K"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Resend Corrected Project Endpoint",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/api/v1/pipeline/{order_id}/start": {
      "post": {
        "description": "Запустить обработку заявки (после загрузки ТУ).\n\nВызывается лендингом после того, как клиент:\n1. Заполнил форму → POST /api/v1/orders\n2. Загрузил ТУ → POST /api/v1/orders/{id}/files?category=tu\n3. Нажал «Отправить» → POST /api/v1/pipeline/{id}/start",
        "operationId": "start_pipeline_api_v1_pipeline__order_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "_k",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": " K"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Pipeline",
        "tags": [
          "pipeline"
        ]
      }
    },
...
  }
}